Add valid URI checks in mapping service graph (#45)
Bioregistry has some synonyms that contain spaces, which currently get
propagated to the `curies.Converter` instance in the Bioregistry's
default `bioregistry.Manager`. This is an issue since it results in
`https://bioregistry.io/<prefix>:<identifier>` variant URIs with spaces.

This PR makes sure that the ones containing spaces get filtered out. It
also does some refactoring of the logic for the custom `triples()`
function to streamline it and reduce code duplication.