Add reference datatypes and switch to pydantic (#58)
In many projects, I have had to re-define data types for parsed CURIEs.
since this is the most fundamental package for handling these objects,
it makes sense to put two flavors here - one as a named tuple and one as
a pydantic class. Basically, they both have the same interface but are
useful in different places.