Add subsetting functionality (#67)
This functionality is useful for downstream applications like the
following:

1. You load a comprehensive extended prefix map, e.g., from the
Bioregistry using `curies.get_bioregistry_converter()`.
2. You load some data that conforms to this prefix map by convention.
This is often the case for semantic mappings stored in the SSSOM format
3. You extract the list of prefixes _actually_ used within your data
4. You subset the detailed extended prefix map to only include prefixes
relevant for your data
5. You make some kind of output of the subsetted extended prefix map to
go with your data. Effectively, this is a way of reconciling data. This
is especially effective when using the Bioregistry or other
comprehensive extended prefix maps.

Here's a concrete example of doing this (which also includes a bit of
data science)
to do this on the SSSOM mappings from the Disease Ontology project.

```python
>>> import curies
>>> import pandas as pd
>>> import itertools as itt
>>> commit = "faca4fc335f9a61902b9c47a1facd52a0d3d2f8b"
>>> url = f"https://github.com/mapping-commons/disease-mappings/blob/{commit}/mappings/doid.sssom.tsv"
>>> df = pd.read_csv(url, sep="\t", comment="#")
>>> prefixes = {
...     curies.Reference.from_curie(curie).prefix
...     for column in ["subject_id", "predicate_id", "object_id"]
...     for curie in df[column]
... }
>>> converter = curies.get_bioregistry_converter()
>>> slim_converter = converter.get_subconverter(prefixes)
```


This PR also sneaks in a related documentation update to pandas
dataframe processing