Add converter with configurable pre-processing (#171)
This PR adds an extension to the `curies.Converter` that allows for
pre-configuring string processing.

This is necessary in many places where what;s possible with simple
contraction and expansion isn't enough to parse CURIEs, URIs, or other
strings that might appear in places where CURIEs or URIs are supposed to
be.

The idea and draft code for this PR existed already in PyOBO, but this
PR generalizes and makes it fully reusable.